shopco-orders-service/requirements.txt
psycopg2-binary
SQLAlchemy==1.3.5